+++ src/main/scala/edu/ie3/simona/service/weather/WeatherService.scala
@@ -26,7 +26,7 @@ import edu.ie3.simona.service.SimonaService
 import edu.ie3.simona.service.weather.WeatherSource.WeightedCoordinates
 import edu.ie3.simona.util.TickUtil.RichZonedDateTime
 import edu.ie3.simona.util.{Coordinate, SimonaConstants}
-import edu.ie3.util.scala.collection.immutable.SortedDistinctSeq
+import edu.ie3.util.scala.collection.immutable.ActivationTickQueue
 import org.apache.pekko.actor.typed.ActorRef
 import org.apache.pekko.actor.typed.scaladsl.ActorContext
 import org.slf4j.Logger
@@ -78,11 +78,8 @@ object WeatherService extends SimonaService {
     *   The weather source to retrieve information from.
     * @param coordinateData
     *   A map of the requested coords to their receiving actor references.
-    * @param maybeNextActivationTick
-    *   The next tick at which the service wants to be activated.
     * @param activationTicks
-    *   A sorted set of ticks, that yet have been sent to the scheduler (w\o
-    *   next tick).
+    *   A queue of future ticks that the service will be activated at.
     * @param startDateTime
     *   The simulation time at which simulation started.
     * @param amountOfInterpolationCoords
@@ -91,8 +88,7 @@ object WeatherService extends SimonaService {
   final case class WeatherBaseStateData(
       weatherSource: WeatherSource,
       coordinateData: Map[Coordinate, CoordinateData] = Map.empty,
-      maybeNextActivationTick: Option[Long],
-      activationTicks: SortedDistinctSeq[Long] = SortedDistinctSeq.empty,
+      activationTicks: ActivationTickQueue = ActivationTickQueue.empty,
       startDateTime: ZonedDateTime,
       amountOfInterpolationCoords: Int = 4,
   ) extends ServiceBaseStateData
@@ -122,25 +118,24 @@ object WeatherService extends SimonaService {
         val weatherSource = WeatherSource(sourceDefinition)
 
         /* What is the first tick to be triggered for? And what are further activation ticks */
-        val (maybeNextTick, furtherActivationTicks) = SortedDistinctSeq(
+        val activationTicks = ActivationTickQueue(
           weatherSource
             .getDataTicks(
               SimonaConstants.FIRST_TICK_IN_SIMULATION,
               simulationEnd.toTick,
             )
             .toSeq
-        ).pop
+        )
 
         val weatherInitializedStateData = WeatherBaseStateData(
           weatherSource,
-          activationTicks = furtherActivationTicks,
-          maybeNextActivationTick = maybeNextTick,
+          activationTicks = activationTicks,
           startDateTime = startDateTime,
         )
 
         Success(
           weatherInitializedStateData,
-          maybeNextTick,
+          activationTicks.nextTick,
         )
 
       case invalidData =>
@@ -209,7 +204,7 @@ object WeatherService extends SimonaService {
     )
 
     // collate the provided coordinates into a single entity
-    val registrationResponse = serviceStateData.maybeNextActivationTick
+    val registrationResponse = serviceStateData.activationTicks.nextTick
       .map(RegistrationSuccessfulMessage(ctx.self, _))
       .getOrElse(RegistrationFailedMessage(ctx.self))
 
@@ -275,13 +270,10 @@ object WeatherService extends SimonaService {
     given simulationStart: ZonedDateTime = serviceStateData.startDateTime
 
     /* Pop the next activation tick and update the state data */
-    val (
-      maybeNextTick: Option[Long],
-      updatedStateData: WeatherBaseStateData,
-    ) = {
-      val (nextTick, remainderTicks) = serviceStateData.activationTicks.pop
-      (nextTick, serviceStateData.copy(activationTicks = remainderTicks))
-    }
+    val remainingTicks = serviceStateData.activationTicks.dropFirst
+    val maybeNextTick = remainingTicks.nextTick
+    val updatedStateData =
+      serviceStateData.copy(activationTicks = remainingTicks)
 
     // get the weather and send it to the subscribed agents
     // no sanity check needed here as we can assume that we always have weather available